add 307 prod variant
