+++ app/src/main/java/com/leetcode_kotlin/TestingData.kt
@@ -109,17 +109,6 @@ fun ListNode.toList(): List<Int> {
     return res
 }
 
-fun IntArray.toListNode(): ListNode {
-    if (this.isEmpty()) return ListNode(0)
-    var stub = ListNode(this[0])
-    val head = stub
-    for (i in 1 until this.size) {
-        val temp = ListNode(this[i])
-        stub.next = temp
-        stub = stub.next!!
-    }
-    return head
-}
 
 object Node {
